+++ spec/rspec/core/memoized_helpers_spec.rb
@@ -254,6 +254,21 @@ module RSpec::Core
           expect(subject_value).to eq(:inner)
         end
 
+        it 'is not overriden when an inner group defines a new method with the same name' do
+          subject_value = nil
+
+          ExampleGroup.describe do
+            subject(:named) { :outer_subject }
+
+            describe "inner" do
+              let(:named) { :inner_named }
+              example { subject_value = self.subject }
+            end
+          end.run
+
+          expect(subject_value).to be(:outer_subject)
+        end
+
         context 'when `super` is used' do
           def should_raise_not_supported_error(&block)
             ex = nil