+++ Changelog.md
@@ -10,6 +10,12 @@ Enhancements
 * Add some logic to test time duration precision. Make it a
   function of time, dropping precision as the time increases. (Aaron Kromer)
 
+Bug fixes
+
+* Fix named subjects so that if an inner group defines a method that
+  overrides the named method, `subject` still retains the originally
+  declared value (Myron Marston).
+
 ### 2.13.1 / 2013-03-12
 [full changelog](http://github.com/rspec/rspec-core/compare/v2.13.0...v2.13.1)